feat(): add navigation and route params
