+++ navigators/ingredientsStack.js
@@ -2,7 +2,6 @@ import React from 'react';
 import { createStackNavigator } from '@react-navigation/stack';
 
 import IngredientsScreen from '../screens/Ingredients/IngredientsScreen';
-import IngredientScreen from '../screens/Ingredients/IngredientScreen';
 import FormIngredientScreen from '../screens/Ingredients/FormIngredientScreen';
 
 const MainIngredientsStack = createStackNavigator();
@@ -10,9 +9,14 @@ const MainIngredientsStack = createStackNavigator();
 function IngredientsStackScreen() {
   return (
     <MainIngredientsStack.Navigator initialRouteName="Ingredients" >
-      <MainIngredientsStack.Screen name="Ingredients" component={IngredientsScreen} />
-      <MainIngredientsStack.Screen name="FormIngredient" component={FormIngredientScreen} />
-      <MainIngredientsStack.Screen name="OneIngredient" component={IngredientScreen} />
+      <MainIngredientsStack.Screen
+        name="Ingredients"
+        component={IngredientsScreen}
+      />
+      <MainIngredientsStack.Screen
+        name="FormIngredient"
+        component={FormIngredientScreen}
+      />
     </MainIngredientsStack.Navigator>
   );
 }
