+++ screens/Ingredients/FormIngredientScreen.js
@@ -13,65 +13,14 @@ import RNPickerSelect from 'react-native-picker-select';
 import { useStoreActions } from 'easy-peasy';
 
 import styles from '../../styles/Ingredients/formStyles';
+import customPickerStyles from '../../styles/customPickerStyles';
 
-const customPickerStyles = {
-  inputWeb: {
-    color: '#AAAAAA',
-    paddingTop: 13,
-    paddingHorizontal: 10,
-    paddingBottom: 12,
-    borderColor: '#AAAAAA',
-    borderWidth: 2,
-    borderRadius: 5,
-    fontSize: 16,
-    paddingLeft: 15,
-    paddingRight: 15,
-    fontWeight: '800',
-    fontStyle: 'normal',
-  },
-  inputIOS: {
-    color: '#111111',
-    paddingTop: 13,
-    paddingHorizontal: 10,
-    paddingBottom: 12,
-  },
-  inputAndroid: {
-    color: '#111111',
-    paddingTop: 13,
-    paddingHorizontal: 10,
-    paddingBottom: 12,
-    width: '100%',
-    height: '60%',
-    borderColor: '#AAAAAA',
-    borderWidth: 2,
-    borderRadius: 5,
-    fontSize: 16,
-  },
-  placeholderColor: '#111111',
-  underline: { borderTopWidth: 0 },
-  icon: {
-    position: 'absolute',
-    backgroundColor: 'transparent',
-    borderTopWidth: 5,
-    borderTopColor: '#00000099',
-    borderRightWidth: 5,
-    borderRightColor: 'transparent',
-    borderLeftWidth: 5,
-    borderLeftColor: 'transparent',
-    width: 0,
-    height: 0,
-    top: 20,
-    right: 15,
-  },
-};
-
-function FormIngredient(props) {
+function FormIngredient({ navigation, route }) {
   const {
     isNew,
-    setShowNewIngredient = null,
-    setShowEditIngredient = null,
     ingredient = null,
-  } = props;
+  } = route.params;
+
   const createIngredient = useStoreActions((actions) => actions.createIngredient);
   const editIngredient = useStoreActions((actions) => actions.createIngredient);
 
@@ -99,7 +48,7 @@ function FormIngredient(props) {
       };
       createIngredient(body)
         .then(() => {
-          setShowNewIngredient(false);
+          navigation.navigate('Ingredients');
         })
         .catch((err) => {
           setShowError(true);
@@ -126,7 +75,7 @@ function FormIngredient(props) {
       };
       editIngredient(body)
         .then(() => {
-          setShowEditIngredient(false);
+          navigation.navigate('Ingredients', { fromEdit: true });
         })
         .catch((err) => {
           setShowError(true);
@@ -146,11 +95,6 @@ function FormIngredient(props) {
 
   return (
     <View style={styles.container}>
-      <View style={[styles.titleRow, styles.title]}>
-        <Text style={styles.rowTitle}>
-          {isNew ? 'Nuevo ingrediente' : 'Editar ingrediente'}
-        </Text>
-      </View>
       <View style={styles.inputContainer}>
         <Text style={styles.inputLabel}>
           Nombre
@@ -210,7 +154,9 @@ function FormIngredient(props) {
       <View style={styles.buttonsContainer}>
         <TouchableOpacity
           style={[styles.button, styles.cancel]}
-          onPress={isNew ? () => setShowNewIngredient(false) : setShowEditIngredient(false)}
+          onPress={isNew ?
+            () => navigation.navigate('Ingredients') :
+            navigation.navigate('Ingredients', { fromEdit: true })}
         >
           <Text style={[styles.buttonText, styles.cancelText]}>
             Cancelar
